\name{augment_nodelist}
\alias{augment_nodelist}

\title{
Augment a node list with graph-based node attributes
}
\description{
\code{augment_nodelist()} augments the supplied node list by adding graph-based measures as node attributes.
}
\usage{
augment_nodelist(nodelist, edgelist = NULL, graph = NULL)
}
\arguments{
  \item{nodelist}{
a \code{"netlit_nodelist"} object, the output of a call to \code{\link{make_nodelist}}.
}
  \item{edgelist}{
optional; an \code{"netlit_edgelist"} object, the output of a call to \code{\link{make_edgelist}}.
}
  \item{graph}{
an \code{igraph} object; the output of a call to \code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame}(edgelist)}.
}
}
\details{
When \code{edgelist} or \code{graph} are supplied, the \code{"degree"} node attribute is added, resulting from \code{\link[igraph:degree]{igraph::degree}(graph, mode = "in")}. (If \code{graph} is not supplied but \code{edgelist} is, \code{igraph::graph_from_data_frame()} will be run on \code{edgelist} first.) In any case, the \code{"betweenness"} attribute is added, resulting from \code{\link[igraph:degree]{igraph::betweenness}}.
}
\value{
The argument supplied to \code{ndoelist} with one or two columns added. When \code{edgelist} or \code{graph} are supplied, the \code{"degree"} column is added. In any case, the \code{"betweenness"} column is added.
}

\seealso{
\code{\link{review}}, \code{\link{make_edgelist}}, \code{\link{make_nodelist}}
}
\examples{
data("literature")
data("node_attributes")

nodelist <- make_nodelist(literature,
                          node_attributes = node_attributes)

edgelist <- make_edgelist(literature,
                          edge_attributes = c("cites", "cites_empirical"))

nodelist_augmented <- augment_nodelist(nodelist,
                                       edgelist = edgelist)

str(nodelist_augmented)

\dontrun{
#Same results:
graph <- igraph::graph_frm_data_from(edgelist)
nodelist_augmented <- augment_nodelist(nodelist,
                                       graph = graph)
}
}