% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_input.R
\name{clean_input}
\alias{clean_input}
\title{Clean input data}
\usage{
clean_input(data)
}
\arguments{
\item{data}{is a dataframe with columns "from" and "to" as well as edge attributes (optional).}
}
\description{
This function cleans the input.
}
\examples{
clean_input()
}
\keyword{clean}
