% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Studies of redistricting.}
\format{
A data frame with x rows and x variables:
\describe{
\item{to}{the to node}
\item{from}{the from node}
...
}
}
\source{
\url{https://github.com/judgelord/redistricting-lit}
}
\usage{
data
}
\description{
A dataset containing the relationships identified in the redistricting literature.
}
\keyword{datasets}
