\name{literature}
\alias{literature}
\docType{data}
\title{
Studies on redistricting
}
\description{
A dataset containing the relationships between concepts identified in the redistricting literature.
}
\usage{data("literature")}
\format{
  A data frame with 69 observations on the following 4 variables.
  \describe{
    \item{\code{from}}{a character vector}
    \item{\code{to}}{a character vector}
    \item{\code{cites}}{a character vector}
    \item{\code{cites_empirical}}{a character vector}
  }
}
\source{
\url{https://github.com/judgelord/redistricting-lit}
}
\examples{
data(literature)
}
\keyword{datasets}
