% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_edgelist.R
\name{make_edgelist}
\alias{make_edgelist}
\title{Make edgelist}
\usage{
make_edgelist(data, edge_attributes = "")
}
\arguments{
\item{data}{is a dataframe with columns "from" and "to" as well as edge attributes (optional). Edge attribute names may not be the same as any node attribute names.}

\item{edge_attributes}{(optional) is a vector of column names for any edge attributes. Edge attribute names may not be the same as any node attribute names.}
}
\description{
This function makes edge list from the input (edgelist) data frame.
}
\examples{
make_edgelist()
}
\keyword{edge}
