\name{make_nodelist}
\alias{make_nodelist}
\title{Make nodelist}

\description{
\code{make_nodelist()} makes a node list from the the "to" and "from" columns of the supplied data frame. A node list is a data frame with a "node" column and, optionally, columns for node attributes, where each row is a node.
}

\usage{
make_nodelist(data, node_attributes = NULL,
              from = "from", to = "to",
              node = "node")
}
\arguments{
\item{data}{
a data set with a "from" column and a "to" column, where each row represents a directed edge. (See the \code{from} and \code{to} arguments below if these columns have names other than \code{"from"} and \code{"to"}.)
}
\item{node_attributes}{
an optional data frame with a "node" column and columns for node attributes, where each row represents a node. (See the \code{node} argument below if this column has a name other than \code{"node"}.)
}
\item{from}{
the name of the column in \code{data} that should be considered the "from" column, ie., representing the nodes from which directed edges emanate. Default is \code{"from"}.
}
\item{to}{
the name of the column in \code{data} that should be considered the "to" column, ie., representing the nodes to which directed edges point. Default is \code{"to"}.
}
\item{node}{
the name of the column in \code{node_attributes} that should be considered the "node" column that contains the names of the nodes. Default is \code{"node"}.
}
}
\value{
A \code{netlit_nodelist} object, which is a data frame with a node column containing the names of all of the nodes present in \code{data}. If \code{node_attributes} is supplied, additional columns for node attributes will be present. The \code{"node"} attribute contains the name of the "node" column in the returned object, which will always be the first column.
}
\seealso{
\code{\link{review}}, \code{\link{augment_nodelist}}
}
\examples{
data("literature")
data("node_attributes")

nodelist <- make_nodelist(literature,
                          node_attributes = node_attributes)

str(nodelist)
}