\name{node_attributes}
\alias{node_attributes}
\docType{data}
\title{
Attributes of redistricting concepts
}
\description{
A dataset with the "type"" of each concept named in the "to" and "from" columns of the \code{\link{literature}} dataset.
}
\usage{data("node_attributes")}
\format{
  A data frame with 108 observations on the following 2 variables.
  \describe{
    \item{\code{node}}{a character vector}
    \item{\code{type}}{a character vector}
  }
}
\examples{
data("node_attributes")
}
\keyword{datasets}
