\name{review}
\alias{review}
\title{Create an edgelist and nodelist augmented with \code{igraph} network statistics}
\description{
Processes a dataframe (and edgelist) by adding network statistics and producing a nodelist and an \code{igraph} graph object for use in subsequent analysis.
}
\usage{
review(data, edge_attributes = NULL,
       node_attributes = NULL, from = "from",
       to = "to", node = "node")
}
\arguments{
\item{data}{
a data set with a "from" column and a "to" column, where each row represents a directed edge. (See the \code{from} and \code{to} arguments below if these columns have names other than \code{"from"} and \code{"to"}.)
}
\item{edge_attributes}{
an optional character vector of the names of columns in \code{data} containing edge attributes. Specified columns will be retained in the augmented edgelist. Others will be dropped. See \code{\link{make_edgelist}}.
}
\item{node_attributes}{
an optional data frame with columns "node" as well as any node attributes. Node attribute names may not be the same as any edge attribute names. See \code{\link{make_nodelist}}
}
\item{from}{
the name of the column in \code{data} that should be considered the "from" column, i.e., representing the nodes from which directed edges emanate. Default is \code{"from"}.
}
\item{to}{
the name of the column in \code{data} that should be considered the "to" column, i.e., representing the nodes to which directed edges point. Default is \code{"to"}.
}
\item{node}{
the name of the column in \code{node_attributes} that should be considered the "node" column that contains the names of the nodes. These names should correspond to node names in the "to" or "from" columns of the main data. Default is \code{"node"}.
}
}
\value{
A \code{netlit_review} object, which contains three elements:
\item{edgelist}{an edge list with the selected edge attributes and the \code{edge_betweenness} column added}
\item{nodelist}{a node list with the included node attributes and the \code{degree_in}, \code{degree_out},  \code{degree_total} and \code{betweenness} columns added}
\item{graph}{an \code{igraph} graph object created from the \code{edgelist} and \code{nodelist}}

There is a \code{print} method that reports some basic summaries of the object.
}
\details{
\code{review()} performs the tasks of \code{\link{make_edgelist}}, \code{\link{make_nodelist}}, and \code{\link{augment_nodelist}} all in one step. First, it creates a pared down version of \code{data} containing only the \code{from}, \code{to}, and \code{edge_attributes} columns, and an added \code{"edge_betweenness"} column containing edge betweennesses as produced by \code{\link[igraph:edge_betweenness]{igraph::edge_betweenness()}}. Next, it creates a node list by merging the \code{attrbiutes} with the unique nodes present in \code{data}. Next, it augments the node list by adding a \code{"degree"} column resulting from \code{\link[igraph:degree]{igraph::degree}} with \code{mode = "in"} and a \code{"betweenness"} column resulting from \code{\link[igraph:degree]{igraph::betweenness}}. Finally, it produces a graph from the edge list and node list using \code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame}}.

All but the last step can be performed individually using the \pkg{netlit} functions mentioned above.
}
\seealso{
\code{\link{make_edgelist}}, \code{\link{make_nodelist}}, \code{\link{augment_nodelist}}

\code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame}}
}
\examples{
data("literature")
lit <- review(literature)
lit

data("node_attributes")
lit <- review(literature,
              edge_attributes = c("cites", "cites_empirical"),
              node_attributes = node_attributes)
lit
}
