% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_nodelist.R
\name{augment_nodelist}
\alias{augment_nodelist}
\title{Augment nodelist and edgelist with igraph network statisitics}
\usage{
augment_nodelist(nodelist, edgelist)
}
\arguments{
\item{nodelist}{is a dataframe with columns "node" as well as any node attributes. Node attribute names may not be the same as any edge attribute names.}

\item{edgelist}{is a dataframe with columns "from" and "to" as well as edge attributes (optional). Edge attribute names may not be the same as any node attribute names.}
}
\description{
This function augments the node list with outputs from the igraph package.
}
\examples{
augment_nodelist()
}
\keyword{node}
