\name{make_edgelist}
\alias{make_edgelist}
\title{Make edgelist}
\description{
\code{make_edgelist} makes an edge list from the supplied data frame. An edge list is a data frame with a "from" column, a "to" column, and, optionally, columns containing edge attributes, where each row is a directed edge.
}
\usage{
make_edgelist(data, edge_attributes = NULL,
              from = "from", to = "to")
}
\arguments{
\item{data}{
a data set with a "from" column and a "to" column, where each row represents a directed edge. (See the \code{from} and \code{to} arguments below if these columns have names other than \code{"from"} and \code{"to"}.)
}
\item{edge_attributes}{
an optional character vector of the names of columns in \code{data} containing edge attributes.
}
\item{from}{
the name of the column in \code{data} that should be considered the "from" column, i.e., representing the nodes from which directed edges emanate. Default is \code{"from"}.
}
\item{to}{
the name of the column in \code{data} that should be considered the "to" column, i.e., representing the nodes to which directed edges point. Default is \code{"to"}.
}
}
\value{
A \code{netlit_edglist} object, which is a data frame identical to \code{data} except for the following:
\itemize{
\item all missing values will have been removed (with a warning if present in \code{data})
\item only the \code{from}, \code{to}, and \code{edge_attributes} columns will be retained, with the \code{from} and \code{to} columns as the first two columns
\item an added column \code{"edge_betweenness"} will be present containing the edge betweenness for each edge
\item the attribute \code{"edge_attributes"} will contain the values supplied to \code{edge_attributes}, if any
}
}
\details{
\code{make_edgelist()} creates a pared down version of \code{data} containing only the \code{from}, \code{to}, and \code{edge_attributes} columns, and an added \code{"edge_betweenness"} column containing edge betweennesses as produced by \code{\link[igraph:edge_betweenness]{igraph::edge_betweenness()}}.
}
\seealso{
\code{\link{review}}
}
\examples{
data("literature")

edgelist <- make_edgelist(literature,
                          edge_attributes = c("cites", "cites_empirical"))

str(edgelist)
}
