% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nodelist.R
\name{make_nodelist}
\alias{make_nodelist}
\title{Make nodelist}
\usage{
make_nodelist(data, node_attributes = tibble(node = ""))
}
\arguments{
\item{data}{is a dataframe with columns "from" and "to" as well as edge attributes (optional). Edge attribute names may not be the same as any node attribute names.}

\item{node_attributes}{(optional) is a dataframe with columns "node" as well as node attributes. Node attribute names may not be the same as any edge attribute names.}
}
\description{
This function makes a node list from the input (edgelist) data frame.
}
\keyword{node}
