% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{Creates a nodelist augmented with igraph network statisitics}
\usage{
review(data, edge_attributes = "", node_attributes = tibble(node = ""))
}
\arguments{
\item{node_attributes}{(optional) is a dataframe with columns "node" as well as any node attributes. Node attribute names may not be the same as any edge attribute names.}

\item{edgelist}{is a dataframe with columns "from" and "to" as well as edge attributes (optional). Edge attribute names may not be the same as any node attribute names.}
}
\value{
A list of 2: an edgelist and a nodelist
}
\description{
This function creates a node list with outputs from the igraph package.
}
\examples{
review(data)
}
\keyword{node}
